reorganize directory structure (move tests into the files they test)
