+++ trace-normalization/src/normalize_utils.rs
@@ -173,3 +173,56 @@ pub fn is_alpha(c: char) -> bool {
 pub fn is_alpha_num(c: char) -> bool {
     is_alpha(c) || ('0'..='9').contains(&c)
 }
+
+#[cfg(test)]
+mod tests {
+
+    use crate::normalize_utils;
+
+    #[test]
+    fn test_normalize_name() {
+        let test_tuples: [(&str, &str, &str); 5] = [
+            (
+                "",
+                "",
+                "Normalizer Error: Empty"
+            ),
+            (
+                "good",
+                "good",
+                "",
+            ),
+            (
+                "Too-Long-.Too-Long-.Too-Long-.Too-Long-.Too-Long-.Too-Long-.Too-Long-.Too-Long-.Too-Long-.Too-Long-.Too-Long-.",
+                "Too_Long.Too_Long.Too_Long.Too_Long.Too_Long.Too_Long.Too_Long.Too_Long.Too_Long.Too_Long.",
+                "",
+            ),
+            (
+                "bad-name",
+                "bad_name",
+                "",
+            ),
+            (
+                "&",
+                "",
+                "Normalizer Error: Invalid"
+            )
+        ];
+
+        for tuple in test_tuples.iter() {
+            let input = tuple.0;
+            let expected = tuple.1;
+            let expected_err = tuple.2;
+
+            match normalize_utils::normalize_name(input.to_string()) {
+                Ok(val) => {
+                    assert_eq!(expected_err, "");
+                    assert_eq!(val, expected);
+                },
+                Err(err) => {
+                    assert_eq!(format!("{}", err), expected_err);
+                }
+            }
+        }
+    }
+}
