+++ trace-normalization/src/pb/span.proto
